feat: drop python 3.6 support
BREAKING CHANGE: Drop Python 3.6 support due to end of life.
