+++ .github/workflows/cicd.yml
@@ -17,12 +17,9 @@ jobs:
     runs-on: ubuntu-latest
     if: "!contains(github.event.head_commit.message, '[skip ci]')"
     steps:
-      - uses: actions/checkout@v2
-      - uses: ./.github/actions/detect-env
+      - uses: actions/checkout@v3.0.2
       - name: Setup Environment
-        uses: ./.github/actions/python
-        with:
-          python-version: ${{ env.PYTHON_VERSION }}
+        uses: ./.github/actions/setup-env
       - name: Install project dependencies
         run: . script/bootstrap
       - name: Lint
@@ -35,23 +32,28 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest]
-        python-version: ['3.6', '3.7', '3.8', '3.9', '3.10']
+        python-version: ['3.7', '3.8', '3.9', '3.10', '3.11-dev']
       fail-fast: false
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3.0.2
       - name: Setup Environment
-        uses: ./.github/actions/python
+        uses: ./.github/actions/setup-env
         with:
           python-version: ${{ matrix.python-version }}
       - name: Install project dependencies
         run: . script/bootstrap
         shell: bash
-      - name: Run tests
+      - name: Run Tests
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         shell: bash
-        run: |
-          poetry run invoke test --coverage
+        if: matrix.python-version != '3.11-dev'
+        run: poetry run invoke test --coverage
+      # pyinvoke doesn't support python 3.11 yet
+      - name: Run Tests (skip invoke)
+        shell: bash
+        if: matrix.python-version == '3.11-dev'
+        run: poetry run pytest ./tests
 # # TODO: How to do this with poetry?
   # version_checks:
   #   name: Dependency Version Constraint Checks
@@ -60,11 +62,10 @@ jobs:
   #   env:
   #     SKIP_DEPS: 1
   #   steps:
-  #     - uses: actions/checkout@v2
-  #     - uses: ./.github/actions/detect-env
-  #     - name: Setup Environment
-  #       uses: ./.github/actions/python
-  #       with:
+  #    - uses: actions/checkout@v3.0.2
+  #    - name: Setup Environment
+  #      uses: ./.github/actions/setup-env
+  #      with:
   #         python-version: 3.7 # it's min, so we'll use an older version of python
   #     - name: Install project dependencies
   #       run: |
@@ -82,14 +83,11 @@ jobs:
     # Dry run check does not need to wait for others
     if: (github.event_name == 'pull_request' || github.ref == 'refs/heads/next') && !contains(github.event.head_commit.message, '[skip ci]')
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3.0.2
         with:
           fetch-depth: 0
-      - uses: ./.github/actions/detect-env
       - name: Setup Environment
-        uses: ./.github/actions/python
-        with:
-          python-version: ${{ env.PYTHON_VERSION }}
+        uses: ./.github/actions/setup-env
       - name: Install project dependencies
         run: . script/bootstrap
       - name: Dry Release
@@ -109,15 +107,12 @@ jobs:
     needs: [analysis, tests] # [version_checks]
     if: github.event_name == 'push' && github.ref == 'refs/heads/master' && !contains(github.event.head_commit.message, '[skip ci]')
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3.0.2
         with:
           fetch-depth: 0
           persist-credentials: false
-      - uses: ./.github/actions/detect-env
       - name: Setup Environment
-        uses: ./.github/actions/python
-        with:
-          python-version: ${{ env.PYTHON_VERSION }}
+        uses: ./.github/actions/setup-env
       - uses: cycjimmy/semantic-release-action@v2
         with:
           branch: master
