+++ .gitignore
@@ -9,6 +9,7 @@ __pycache__/
 # Distribution / packaging
 .Python
 .tox
+.poetry
 build/
 develop-eggs/
 dist/
